update log table
